Migrate existing code to Cobra
