+++ go.mod
@@ -1,3 +1,9 @@
 module bphcat
 
 go 1.19
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/spf13/cobra v1.5.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
